Move startup checks and logs to their own cog
This is to avoid needed to use wait_until_guild_available during the setup hook.
